requirements.txt
python-dotenv==0.15.0
async-timeout==3.0.1
discord.py==1.6.0